display website links clearer
