--- components/SocialLinks/SocialLinks.module.scss
@@ -1,13 +1,0 @@
-@import "styles/mixins";
-
-.wrapper {
-  display: flex;
-  gap: 2rem;
-  justify-content: center;
-}
-
-.icons {
-  border: 1px solid;
-  border-radius: 50%;
-  padding: 0.5rem 0.75rem;
-}
