--- components/SocialLinks/index.tsx
@@ -1,1 +1,0 @@
-export { default as SocialLinks } from "./SocialLinks";
